Fixed a problem with packet header_map beyond 32 bit in shift. Added (1ULL << scan->id) to the bitshift operation to allow 64 bit operation. It fixed the problem where packet.hasHeader(ID) would result in false even if header of type ID existed but its ID was beyond 32 bits.

+++ jnetpcap/trunk/src/c/packet_jsmall_scanner.cpp
@@ -429,7 +429,7 @@ void record_header(scan_t *scan) {
 	/*
 	 * Initialize the header entry in our packet header array
 	 */
-	packet->pkt_header_map |= (1 << scan->id);
+	packet->pkt_header_map |= (uint64_t)(1ULL << scan->id);
 	header->hdr_id = scan->id;
 	header->hdr_offset = offset + scan->hdr_prefix;
 	header->hdr_analysis = NULL;
